Start implementing a simple Planner class
Currently, the planner solves trivial problems only. i.e., problems with
goals that can be achieved by calling a single action.

Note that the action effect block must take in the state as an argument
and manipulate its variables. Otherwise, [the block will be just setting
local variables instead of using the dynamically-added accessors of the
state][1]. In effect, the action will simply not affect the state.

[1]: http://stackoverflow.com/questions/4403973/