"Objectify" the plan's string representation
The new representation is better suited for sequence diagrams: instead
of just listing the method names, this new representation contains the
receiver and the parameters, if any, of each method. So, instead of just
`"method"`, the new representation will contain `"object.method(param_1,
..., param_n)"`.

For the time being, the name of the receiver is computed by downcasing
the name of the class to which the method belongs.