Make State a set of objects
This was a major change that broke Planner, so I changed it accordingly
(except for using use cases to set up the initial state). Planner does
not deal with DbC classes anymore as it doesn't need them: objects
knowing their methods and the names of their classes is sufficient at
the current stage.