Dynamically define accessors for DbC objects
An accessor is defined for each DbC instance variable. These accessors
should come in handy when an instance variable is another DbC object as
the accessor will enable conditions to reference the instance variables
of that object.

These accessors must be defined as singleton methods because otherwise,
different instances of DbcObject representing different instances of
different DbC classes will share the same accessors.