Implement deep-copying for DbC objects
I have chosen to override Object#clone instead of adding a new method to
simplify the code in DbcObject (that is, to avoid checking if the object
to be cloned is a DbcObject and calling that new method on it).

For now, there does not seem to be a need for each clone to have its
own, separate DbC methods. So, copies of the same DbcObject share the
elements of the dbc_methods array.