Implement deterministic forward state-search
This deterministic implementation is a recursive depth-first search.
Because the planner cannot compare states to determine if it had visited
them before (no proper implementation of `State#==` yet), and to avoid
infinite loops, previously-called methods are considered inapplicable to
to the current state. This little hack may need revisiting when an
example that invalidates its assumption arises.