Correct an implementation error of the BFS algorithm
In breadth-first search, declaring failure when the planner cannot find
any applicable methods to the current (non-goal) state is wrong because
there cloud be other states in the queue waiting to be checked.
Basically, this current state is a leaf node, and instead of declaring
failure, the planner should move on and check the next state in the
queue (if any).