Start solving the problem of DbcMethod parameters
Due to the closure nature of Ruby blocks, if a DbcMethod's
pre/postcondition refers to a (DbcObject) parameter, it will always
refer to the object defined where the condition block itself is defined,
never its clones that are generated and modified during (deterministic)
planning. The solution is to get hold of the needed parameter by making
the DbcObject ask its state to get it a DbcObject of the required DbC
class.

Currently, the solution will not work in all cases because there is a
problem in cloning DbcObjects: Because of the current recursive
implementation of `DbcObject#clone`, `State#clone` will create multiple
separate clones of the same DbcObject if it is referred to by multiple
DbcObjects.