Count the number of unsatisfied sub-conditions
More accurately, given a hash of object names and condition blocks,
State can count the number of its (DbC) objects that don't satisfy their
associated conditions. This count should be useful for planning as (part
of) a heuristic that measures the distance to a goal state.

`State#satisfy?` was reimplemented using this new count because the two
implementations were very similar.