Replace unstable priority queue with a stable one
The previous priority queue was unstable in the sense that when it had
more than one item of the same priority, it sometimes behaved in FIFO
manner, and other times in LIFO manner. This instability changed the
order of states being tested for goal and ultimately caused the number
of goal tests to fluctuate instead of (deterministically) increase as
the number of noise methods increased.

When the [Double-Ended Priority Queue][1] is used instead, the number of
goal tests behaves as expected.

[1]:https://github.com/akr/depq