Change the structure of plans and add caller info
A generated (successful) plan is now an array of hashes representing a
sequence of method calls, with each hash (method call) containing the
method's name, its caller's name, its parameters' names, and its
receiver's name.

For now, the method's caller is determined using the part-of
associations between the objects in the current state. As advised by Dr.
Ahmed, if there are multiple candidate callers, the planner selects one
at random. If there are none, the planner assumes that the actor of the
use case is the caller.