Improve the heuristics for determining the caller
The underlying assumptions of these heuristics are based on the examples
I have seen so far and on the discussions with Dr. Ahmed, and they
should generally lead to more accurate sequence diagrams.

I'm starting to test private methods directly because setting up each
unit test is becoming a pain but I currently don't have the time to
change the Planner to make it more testable.