Add a `dead` boolean attribute to DbcObjects
If a DbcObject is marked as dead, it means that it shouldn't be consider
as existing in the State yet and that it is more like a blueprint for
object instantiation.

Marking a DbcObject as dead will make its DbcMethods inapplicable from
the Planner's point of view. Together with dependency relationships
(which are yet to be implemented), the Planner should be able to
instantiate objects (read: unmark as dead) when necessary.

I've decided to make the attribute not true by default to avoid having
to mark all of the initially "alive" DbcObjects I already have in the
tests and examples.