Enable Planner to instantiate dependency objects
It is assumed that whenever a DbcMethod `m()` that depends on a
DbcObject is called, it will immediately create the dependency
DbcObject. That is, `m()` will always be followed by a `<<create>>`
method call. The current implementation is a poor man's version of
object instantiation. For the time being, instantiation simply means:

1. Modifying the State by umarking the dependency DbcObject as dead.
2. Constructing the appropriate `<<create>>` method call and adding it
to the current sequence of method calls.