Add the weather station experiment script
Apart from the first method call, the generated plan does not contain
the messages sent from and to the actor because the library doesn't
support actor interactions yet. I thought about using the trick of
adding a DbcObject representing the actor, but it convolutes the
experiment, not to mention that the current sender-selection rules
result in silly messages (e.g. actor `acknowledge()` itself).